apply code editor on page
