+++ components/index.tsx
@@ -3,7 +3,8 @@ export { default as MainNav } from './MainNav';
 export { default as MainFooter } from './MainFooter';
 export { default as UIWrapper } from './UIWrapper';
 export { default as AlertModal } from './AlertModal';
+export { default as CodeEditor } from './CodeEditor';
 export { default as MainPage } from './paths/MainPage';
 export { default as ContributorsPage } from './paths/ContributorsPage';
 export { default as FeaturesPage } from './paths/FeaturesPage';
-export { default as ViewsPage } from './paths/ViewsPage';
+export { default as CodeViewPage } from './paths/CodeViewPage';
